+++ test/datetime_extension_test.dart
@@ -4,6 +4,7 @@ import 'package:test/test.dart';
 void main() {
   test('main', () {
     final DateTime result = DateTime(2020);
+
     expect(result == result.clone, true);
     expect(result.copyWith(day: 2), result.nextDay);
     expect(result.copyWith(year: 2019, month: 12, day: 31), result.previousDay);
@@ -13,6 +14,7 @@ void main() {
   test('equivalence', () {
     final DateTime result1 = DateTime(2020);
     final DateTime result2 = DateTime(2020, 1, 1, 12, 30);
+
     expect(result1.isSameMoment(result2), false);
     expect(result1.isSameYear(result2), true);
     expect(result1.isSameMonth(result2), true);